refact: Replace PatheticMapper with PatheticFacade and add PathfinderFactory
Refactored PatheticMapper into PatheticFacade and introduced a new PathfinderFactory class. PatheticPlugin now uses PatheticFacade and PathfinderFactory for pathfinder instance creation and initialization.
