--- pathetic-mapping/src/main/java/de/metaphoriker/mapping/PatheticMapper.java
@@ -1,57 +1,0 @@
-package de.metaphoriker.mapping;
-
-import lombok.NonNull;
-import lombok.experimental.UtilityClass;
-import org.bukkit.plugin.java.JavaPlugin;
-import de.metaphoriker.Pathetic;
-import de.metaphoriker.api.pathing.Pathfinder;
-import de.metaphoriker.api.pathing.configuration.PathfinderConfiguration;
-import de.metaphoriker.model.pathing.pathfinder.AStarPathfinder;
-import de.metaphoriker.util.ErrorLogger;
-
-/** PatheticMapper is a utility class that maps the Pathetic API to the Pathetic Implementation. */
-@UtilityClass
-public class PatheticMapper {
-
-  /**
-   * @apiNote If Pathetic is not initialized yet but is used anyways, this will cause many things to
-   *     break.
-   * @param javaPlugin the JavaPlugin which initializes the lib
-   * @throws IllegalStateException If an attempt is made to initialize more than once
-   */
-  public void initialize(JavaPlugin javaPlugin) {
-    Pathetic.initialize(javaPlugin);
-  }
-
-  /**
-   * Signals Pathetic to initiate its shutdown process, releasing resources and finalizing
-   * operations. This method should be called when Pathetic is no longer needed or the plugin is
-   * being disabled.
-   */
-  public void shutdown() {
-    Pathetic.shutdown();
-  }
-
-  /**
-   * Instantiates a new pathfinder object.
-   *
-   * @return The {@link Pathfinder} object
-   * @throws IllegalStateException If the lib is not initialized yet
-   */
-  public @NonNull Pathfinder newPathfinder() {
-    return newPathfinder(PathfinderConfiguration.createAsyncConfiguration());
-  }
-
-  /**
-   * Instantiates a new A*-pathfinder.
-   *
-   * @param pathfinderConfiguration - The {@link PathfinderConfiguration}
-   * @return The {@link Pathfinder}
-   * @throws IllegalStateException If the lib is not initialized yet
-   */
-  public @NonNull Pathfinder newPathfinder(PathfinderConfiguration pathfinderConfiguration) {
-    if (Pathetic.isInitialized()) return new AStarPathfinder(pathfinderConfiguration);
-
-    throw ErrorLogger.logFatalError("Pathetic is not initialized yet.");
-  }
-}
